Resolves the two conflicts this branch had accumulated over 149 commits.

runtime-provider.tsx: keep the branch's reloadReadyThreadId and
onInitialHistoryReady parameters, with main's wrapped return type.

attachment.tsx: main moved AttachmentPreviewDialog out to
attachment-preview.tsx, so take main's file and re-apply the branch's
redaction there instead: the redactFromReload prop now threads through
AttachmentImageDialog onto the lightbox DialogContent, ComposerAttachments
keeps its marker, and the reload-snapshot test points at the new file.

+++ unsloth/models/llama.py
@@ -2478,7 +2478,9 @@ class FastLlamaModel:
         model_function = MODEL_FOR_CAUSAL_LM_MAPPING[model_config.__class__]
         IS_FALCON_H1 = model_config.model_type.startswith("falcon_h1")
 
-        preferred_attn_impl = resolve_attention_implementation(model_function, model_config)
+        preferred_attn_impl = resolve_attention_implementation(
+            model_function, model_config, dtype = dtype
+        )
 
         # Prefetch the repo (killable child) so the weight load is a cache hit. Runs after the
         # AutoConfig/model-class check so an unsupported repo fails on its small config fetch.
